refactor: extract config and tunnel
